Small bug fix in TypeEnvironmentCalculator. More testing for InstertExplicitTypeCasts.

+++ tests/test_transforms/fixtures/insert_explicit_type_casts/01_floats_golden.c
@@ -6,7 +6,7 @@ int main()
   a = (float) (((long double) b) + c);
   b = (double) (((long double) a) + c);
   c = (long double) (b + ((double) a));
-  a = (float) (a + a);
+  a = a + a;
   a = (float) (((long double) (((double) a) + b)) + (c + ((long double) a)));
 
   int i;
